Fixed bug with tracefile 2.0 and refactoring
The original implementation did not include an option where an object
can be allocated to two rootsets. This was not an issue with tracefile
1.0, but tracefile 2.0 allows the allocation of an object ('a') and the
addition to a rootset ('+'). However, the '+' command can happen more
often, meaning that the very same object can be part of several
rootsets. In order to overcome this issue I test for every '+' command
whether we already have this object in our object list. That way the GC
can safely delete it if the object is not alive anymore.

Furthermore, I started refactoring the code to make it more human
readable and to remove old and unused code parts. Fixed the file
extensions (.h -> .hpp, because C++) and changed the last remains from
the programs internal development name to the public name.