Performance Enhancement. Smarter object lookup.
In our tracefile format, objects are identified by their unique ID. To locate
the object in memory, a mapping must be performed from <object ID> to <object
pointer>. These mappings are store in the objectList vector.

The lookup was being naively implemented as a linear search of the vector (each
element was checked). This has been updated, so that each object is stored
according to its ID, allowing for constant time lookup.