Performance Enhancement. Smart Rootset Lookup.
When an object was added to the rootset, it was being naively inserted into a
rootset vector. This meant that to locate a root (either to get its address, or
to test for membership), a linear search was performed.

To increase performance, roots are hashed into a rootset map by their ID. This
means that roots can be looked up in constant time, and also removes the concept
of a 'root-slot' from our simulator.

Now that the concept of 'root-slots' has disappeared, a new method, getRoots(),
has been added to allow the caller to traverse the rootset. getRoots() returns a
vector of all roots (the object pointers) currently in the rootset.