Enhancement. Smarter (yet again) object lookup.
The previous solution to the linear search object lookup bottleneck was solved
by storing objects at the position corresponding to their ID in the objectList
vector. This allowed for constant time lookup, but had a negative effect as
program size grew. Because object IDs are not reused even after an object has
been collected, in a large program, the beginning of the object list becomes
essentially wasted space. This becomes a serious issue when iterating the set of
objects (which happens in a GC) because we must iterate all these 'dead' spaces
as well.

Re-implemented the objectList as a hash-map (named objectMap) to allow constant
time lookup, while avoiding the large number of unnecessary 'dead' slots.

In fact, since live objects are no longer stored in a vector, the concept of a
'list slot' is no longer valid. The ObjectContainer::getbySlotNr() method has
benn removed, and replaced with a getLiveObjects() method which returns a vector
of object pointers. This leads to easier iteration of live objects and also
creates a more uniform interface to the ObjectContainer class.