Refactor. Changed unnecessarily large int fields to bool.
The flags Object::isVisited and Object::isAlive were unnecessarily occupying the
full size of an int. This becomes a problem because when using the real
allocation policy, each object is written to the heap. Theoretically, this
should reduce the size of each object by 6 bytes, but after testing the object
size remained constant. This could potentially be due to alignment?