Real Allocator now searches for free memory using Card Tables.
Previously, the RealAllocator::alloate() method was traversing the heap bit-by-bit
when searching for free, contiguous space. This led to poor performance. Implemented
a card table based search to improve allocation performance.

Additionally, the real allocator did not store an objects pointer fields within the
space allocated for that object on heap. The pointer fields were being malloc'd, and were
in fact not stored in our simulated heap at all. Changed this so that when using the real
allocator, an objects pointer fields are stored within the space allocated for the object.