Fix. Object class now stores data on heap.
The 'raw' data contained in each object (its header and pointer fields) are now
physically stored in our simulated heap. The Object class is now essentailly a
container which stores simulation meta-data (i.e. the object id, garbage
collection flags, ...) and a pointer to the data on-heap.

A couple of notes about the new format:

An attempt has been made to conserve the Object API as much as possible. When a
new Object is instantiated, its heap data will transparently be written to the
corrent heap address. Pointer manipulations can also be made transparently.
Calling Object::setPointer() will cause the on-heap data to be manipulated. To
retrieve the address of the heap-data, use Object::getAddress(). Simply using
&object will return the memory address of the Object class, which is probaly not
what you want.

When writing collection policies, you may require the actual memory address
stored in the pointer field. The transparency of these operations hides the heap
manipulations from the rest of the program. However, if you need to manipulate
the actual memory address stored in a pointer field, use
Object::setRawPointerAddress() and Object::getRawPointerAddress(). These methods
have been included to allow certain, low-level, tasks to be performed.

Finally, two distinct methods exist which are related to the Object's size.
	- Object::getHeapSize() returns the number of bytes that this object occupies
		on the heap. This will be equal to the size specified in the tracefile.
	- Object::getPayloadSize() returns the numbers of bytes that the object
		occupies on the heap, which are NOT part of it's header.