Added a method to retrieve live objects in 'heap order'.
Certain GC policies which include a compaction phase require objects to be
traversed in the order in which they physically appear in the heap (otherwise,
moving an object could overwrite another which hasn't been visited yet).

ObjectContainer::getLiveObjectsInHeapOrder() returns a vector of live objects in
the order in which they appear in the heap.