Refactor. Adjusted interface used for remembering allocation points, such as in nextfit.
To allocate in a nextfit (or similar) fashion, the allocator needs to store the
last checked point in the heap. This was being saved as a pointer, which didn't
work well as all of the other internals were being stored as heap indices.

Simplified the interface for explicitly reseting / overriding the nextfit
approach. Allocating using this remembered strategy can cause problems during a
compaction phase (could corrupt data), therefor the caller needs a way to
override the remembered point. Previously, the caller was asked to hand in an
index into the heap (for example, 0), but why should the caller need to know
about implementation. Simplified to have a zero-parameter
Allocator::resetRememberedHeapIndex() method instead.

The protected method, Allocator::allocate() was expecting the
rememberedHeapIndex to be handed in, however, since rememberedHeapIndex is a
class variable, it is already in scope. Simplified the signature of
Allocator::allocate().