Fix. MarkSweep collector now correctly performs compaction.
Previously, during a compaction phase, the object-to-object pointers were not
being updated after objects were moved during the compaction phase. This led to
corrupt pointers.

Additionally, the iteration order of objects during the compaction phase was not
necessarily being performed in 'heap order' (i.e. the order in which the objects
appear in the heap). This meant that it was possible to overwrite or corrupt an
object that hadn't been visited in the iteration yet.