Added support for pointers which appear in static fields.
The lines beginning with 'c' (which represent storing a reference field into a
class object) were previously being ignored by the simulator. This was
problematic, because it allowed for objects to be potentially collected before
they were, in fact, dead (the case where the only remaining pointers to an
object were from class objects).

These 'c' lines are now being processed by the simulator. Since limited
information is given about each class, these static pointers are stored in a Vector
of HashMaps. The static reference with field offset 'j' belonging to class 'i'
is stored in:
	classReferences[i][j];

All static references are being included along wih the rootset when a collection
is triggered.