Further fix to depth first copy
I forgot to reverse the for loop on threads as well, so later
thread's rootsets were getting copied first. I also changed
when objects get counted as 'visited' on depth-first copy, so
when an object is added to the copy stack more than once, the
later additions take priority. Because of the reversed 'for'
loops, later additions are actually earlier in the rootset.