Efficiency change to allocators, removed faulty tests
If a large contiguous chunk of memory is on the heap, but not large enough to
contain an object that is trying to allocate, the allocator would spin its
wheels, incrementing by one every time it tried again. This change adds the
amount of memory space checked to the counter.

Also removed three test cases that always failed. We don't even use the real allocator anyways.